pyproject.toml
[project]
name = "economic-intelligence-mcp"
version = "1.0.0"
description = "External economic intelligence MCP server for Databricks Supervisor Agent"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.12",
    "mcp[cli]>=1.14.0",
    "uvicorn>=0.34.2",
    "databricks-sdk>=0.60.0",
    "pydantic>=2",
    "fastmcp>=2.12.5",
    "httpx>=0.27.0",
]

[project.scripts]
economic-intelligence-mcp = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
